Generalize SQLite database to interface
