+++ db.go
@@ -4,17 +4,10 @@
 package golink
 
 import (
-	"context"
-	"database/sql"
 	_ "embed"
-	"errors"
-	"fmt"
-	"io/fs"
 	"net/url"
 	"strings"
 	"time"
-
-	_ "modernc.org/sqlite"
 )
 
 // Link is the structure stored for each go short link.
@@ -37,125 +30,10 @@ func linkID(short string) string {
 	return id
 }
 
-// SQLiteDB stores Links in a SQLite database.
-type SQLiteDB struct {
-	db *sql.DB
-}
-
-//go:embed schema.sql
-var sqlSchema string
-
-// NewSQLiteDB returns a new SQLiteDB that stores links in a SQLite database stored at f.
-func NewSQLiteDB(f string) (*SQLiteDB, error) {
-	db, err := sql.Open("sqlite", f)
-	if err != nil {
-		return nil, err
-	}
-	if err := db.Ping(); err != nil {
-		return nil, err
-	}
-
-	if _, err = db.Exec(sqlSchema); err != nil {
-		return nil, err
-	}
-
-	return &SQLiteDB{db: db}, nil
-}
-
-// LoadAll returns all stored Links.
-//
-// The caller owns the returned values.
-func (s *SQLiteDB) LoadAll() ([]*Link, error) {
-	var links []*Link
-	rows, err := s.db.Query("SELECT Short, Long, Created, LastEdit, Owner FROM Links")
-	if err != nil {
-		return nil, err
-	}
-	for rows.Next() {
-		link := new(Link)
-		var created, lastEdit int64
-		err := rows.Scan(&link.Short, &link.Long, &created, &lastEdit, &link.Owner)
-		if err != nil {
-			return nil, err
-		}
-		link.Created = time.Unix(created, 0).UTC()
-		link.LastEdit = time.Unix(lastEdit, 0).UTC()
-		links = append(links, link)
-	}
-	return links, rows.Err()
-}
-
-// Load returns a Link by its short name.
-//
-// It returns fs.ErrNotExist if the link does not exist.
-//
-// The caller owns the returned value.
-func (s *SQLiteDB) Load(short string) (*Link, error) {
-	link := new(Link)
-	var created, lastEdit int64
-	row := s.db.QueryRow("SELECT Short, Long, Created, LastEdit, Owner FROM Links WHERE ID = ?1 LIMIT 1", linkID(short))
-	err := row.Scan(&link.Short, &link.Long, &created, &lastEdit, &link.Owner)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			err = fs.ErrNotExist
-		}
-		return nil, err
-	}
-	link.Created = time.Unix(created, 0).UTC()
-	link.LastEdit = time.Unix(lastEdit, 0).UTC()
-	return link, nil
-}
-
-// Save saves a Link.
-func (s *SQLiteDB) Save(link *Link) error {
-	result, err := s.db.Exec("INSERT OR REPLACE INTO Links (ID, Short, Long, Created, LastEdit, Owner) VALUES (?, ?, ?, ?, ?, ?)", linkID(link.Short), link.Short, link.Long, link.Created.Unix(), link.LastEdit.Unix(), link.Owner)
-	if err != nil {
-		return err
-	}
-	rows, err := result.RowsAffected()
-	if err != nil {
-		return err
-	}
-	if rows != 1 {
-		return fmt.Errorf("expected to affect 1 row, affected %d", rows)
-	}
-	return nil
-}
-
-// LoadStats returns click stats for links.
-func (s *SQLiteDB) LoadStats() (ClickStats, error) {
-	rows, err := s.db.Query("SELECT ID, sum(Clicks) FROM Stats GROUP BY ID")
-	if err != nil {
-		return nil, err
-	}
-	stats := make(map[string]int)
-	for rows.Next() {
-		var id string
-		var clicks int
-		err := rows.Scan(&id, &clicks)
-		if err != nil {
-			return nil, err
-		}
-		stats[id] = clicks
-	}
-	return stats, rows.Err()
-}
-
-// SaveStats records click stats for links.  The provided map includes
-// incremental clicks that have occurred since the last time SaveStats
-// was called.
-func (s *SQLiteDB) SaveStats(stats ClickStats) error {
-	tx, err := s.db.BeginTx(context.TODO(), nil)
-	if err != nil {
-		return err
-	}
-	now := time.Now().Unix()
-	for short, clicks := range stats {
-		_, err := tx.Exec("INSERT INTO Stats (ID, Created, Clicks) VALUES (?, ?, ?)", linkID(short), now, clicks)
-		if err != nil {
-			tx.Rollback()
-			return err
-		}
-	}
-	return tx.Commit()
+type Database interface {
+	LoadAll() ([]*Link, error)
+	Load(short string) (*Link, error)
+	Save(link *Link) error
+	LoadStats() (ClickStats, error)
+	SaveStats(stats ClickStats) error
 }
